backend/requirement.txt
amqp==2.2.1
billiard==3.5.0.3
celery==4.1.0
certifi==2017.7.27.1
chardet==3.0.4
decorator==4.1.2
Django==1.11.3
django-celery-beat==1.0.1
django-cors-headers==2.1.0
django-crispy-forms==1.6.1
django-markdown-deux==1.0.5
django-pagedown==0.1.3
-e git://github.com/darklow/django-suit.git@613c84cba6f8fb6a23b771da96ab8c05808fb5dc#egg=django_suit
djangorestframework==3.6.3
djangorestframework-jwt==1.11.0
flake8==3.4.1
googletrans==2.1.4
h2==2.6.2
hpack==3.0.0
hyper==0.7.0
hyperframe==3.2.0
idna==2.6
ipython==6.1.0
ipython-genutils==0.2.0
jedi==0.10.2
kombu==4.1.0
markdown2==2.3.4
mccabe==0.6.1
munch==2.2.0
pexpect==4.2.1
pickleshare==0.7.4
prompt-toolkit==1.0.15
psycopg2==2.7.3
ptyprocess==0.5.2
pycodestyle==2.3.1
pyflakes==1.5.0
Pygments==2.2.0
PyJWT==1.5.2
pytz==2017.2
redis==2.10.5
requests==2.18.4
requests-toolbelt==0.8.0
simplegeneric==0.8.1
six==1.10.0
traitlets==4.3.2
translate==2.0.0
urllib3==1.22
vine==1.1.4
wcwidth==0.1.7
